Update the caching infrastructure to allow for arbitrary new caching systems to be combined into one fn_extra pointer slot. Clean up the 1-d r-tree code a little bit. Make the cache handling code for geometry caching generic so it can be re-used across different internal indexing approaches
git-svn-id: http://svn.osgeo.org/postgis/trunk@9918 b70326c6-7e19-0410-871a-916f4a2858ee

+++ postgis/lwgeom_ogc.c
@@ -833,7 +833,7 @@ Datum LWGEOM_asText(PG_FUNCTION_ARGS)
 	/* Write to WKT and free the geometry */
 	wkt = lwgeom_to_wkt(lwgeom, WKT_ISO, DBL_DIG, &wkt_size);
 	lwgeom_free(lwgeom);
-	POSTGIS_DEBUGF(3, "WKT size = %d, WKT length = %d", wkt_size, strlen(wkt));
+	POSTGIS_DEBUGF(3, "WKT size = %u, WKT length = %u", (unsigned int)wkt_size, (unsigned int)strlen(wkt));
 
 	/* Write to text and free the WKT */
 	result = cstring2text(wkt);
